requirements.txt
git+git://github.com/jmoswalt/django-things#egg=things-20130401.100

# Caching
johnny-cache==1.4
python-memcached==1.48

# Heroku deployment packages
dj-database-url==0.2.1
psycopg2==2.4.5
gunicorn==0.14.5

# S3 storage packages
django-s3-folder-storage==0.1
django-storages==1.1.5

# For Sentry
raven==3.2.1

markdown==2.3.1
